requirements.txt
httpx>=0.27.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
rich>=13.7.0
pytest>=8.0.0
websockets>=12.0

